backend/requirements.txt
Django>=4.2,<5.0



# test 
pytest 
pytest-django


# refactor and formatting
flake8
black


# django rest framework
djangorestframework
markdown
django-filter